+++ chalice/app.py
@@ -540,7 +540,7 @@ class DecoratorAPI(object):
     def route(self, path, **kwargs):
         return self._create_registration_function(
             handler_type='route',
-            name=kwargs.get('name'),
+            name=kwargs.pop('name', None),
             # This looks a little weird taking kwargs as a key,
             # but we want to preserve keep the **kwargs signature
             # in the route decorator.